move http/userInGroups() to os/UserInGroups() as requested.
