+++ go/http/httpbase.go
@@ -24,6 +24,7 @@ import (
 	"github.com/martini-contrib/auth"
 
 	"github.com/github/orchestrator/go/config"
+	"github.com/github/orchestrator/go/os"
 	"github.com/github/orchestrator/go/process"
 )
 
@@ -65,7 +66,7 @@ func isAuthorizedForAction(req *http.Request, user auth.User) bool {
 				}
 			}
 			// check the user's group is one of those listed here
-			if len(config.Config.PowerAuthGroups) > 0 && userInGroups(authUser, config.Config.PowerAuthGroups) {
+			if len(config.Config.PowerAuthGroups) > 0 && os.UserInGroups(authUser, config.Config.PowerAuthGroups) {
 				return true
 			}
 			return false
